feat(launcher): renamed endpoints and introduced MissionControl.validate
